Add Navigation component based navigation
