+++ app/build.gradle
@@ -4,6 +4,9 @@ apply plugin: 'kotlin-android'
 
 apply plugin: 'com.squareup.sqldelight'
 
+apply plugin: "androidx.navigation.safeargs.kotlin"
+
+
 android {
     compileSdkVersion 32
     buildToolsVersion "32.1.0-rc1"
@@ -50,6 +53,13 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     implementation "androidx.core:core-ktx:1.7.0"
+
+    //Android Jetpack
+    implementation "androidx.navigation:navigation-fragment-ktx:2.5.0-alpha03"
+    // For Kotlin use navigation-fragment-ktx
+    implementation "androidx.navigation:navigation-ui-ktx:2.5.0-alpha03"
+    // For Kotlin use navigation-ui-ktx
+
 }
 repositories {
     mavenCentral()
