+++ app/src/main/java/me/elmanss/melate/ui/MainActivity.kt
@@ -1,111 +1,12 @@
 package me.elmanss.melate.ui
 
-import android.content.Intent
-import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import android.widget.Toast
-import androidx.activity.viewModels
-import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.lifecycleScope
-import androidx.recyclerview.widget.DividerItemDecoration
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-import logcat.logcat
+import android.os.Bundle
 import me.elmanss.melate.R
-import me.elmanss.melate.databinding.ActivityMainBinding
-import me.elmanss.melate.ui.custom.util.ItemClickSupport
-import me.elmanss.melate.ui.favs.FavsActivity
 
-
-class MainActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshListener {
-    private val adapter = MainAdapter()
-    private val viewModel: MainViewModel by viewModels()
-    private lateinit var binding: ActivityMainBinding
-
+class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        val root = binding.root
-        setContentView(root)
-        configBinding(binding)
-        Toast.makeText(
-            this,
-            "Para agregar sorteos a tus favoritos, mantén presionado el sorteo de tu elección",
-            Toast.LENGTH_LONG
-        ).show()
-
-        observe()
-    }
-
-    private fun configBinding(binding: ActivityMainBinding) {
-        binding.apply {
-            mainScreen.setOnRefreshListener(this@MainActivity)
-            mainSorteosView.layoutManager = LinearLayoutManager(this@MainActivity)
-            mainSorteosView.addItemDecoration(
-                DividerItemDecoration(
-                    this@MainActivity,
-                    LinearLayoutManager.VERTICAL
-                )
-            )
-
-            mainSorteosView.adapter = adapter
-            ItemClickSupport.addTo(mainSorteosView).setOnItemClickListener { _, pos, _ ->
-                showWarning(pos)
-            }
-        }
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.menu_main, menu)
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.item_favs -> startActivity(Intent(this, FavsActivity::class.java))
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-    private fun observe() {
-        viewModel.sorteos.observe(this) {
-            it?.let {
-                adapter.add(it)
-                logcat { "Added item ${it.prettyPrint()} at position ${adapter.itemCount - 1}" }
-            }
-        }
-    }
-
-    private fun showWarning(pos: Int) {
-        AlertDialog.Builder(this)
-            .setTitle("Aviso")
-            .setMessage("¿Deseas agregar este sorteo de tu lista de favoritos?")
-            .setPositiveButton("Si") { d, _ ->
-                viewModel.saveToFavorites(adapter.getItem(pos))
-                Toast.makeText(
-                    this@MainActivity,
-                    "Sorteo agregado a tus favoritos",
-                    Toast.LENGTH_SHORT
-                )
-                    .show()
-                d.dismiss()
-            }.show()
-    }
-
-    override fun onRefresh() {
-        1500L.launchOnRefresh()
-    }
-
-    private fun Long.launchOnRefresh() {
-        lifecycleScope.launch {
-            adapter.clear()
-            delay(this@launchOnRefresh)
-            binding.root.isRefreshing = false
-            viewModel.fetchSorteos()
-        }
+        setContentView(R.layout.activity_main)
     }
 }
