+++ app/src/main/res/layout/fragment_favs.xml
@@ -6,7 +6,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.favs.FavsActivity">
+        tools:context=".ui.MainActivity">
 
 
     <androidx.recyclerview.widget.RecyclerView
@@ -23,6 +23,7 @@
             android:id="@+id/addButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:backgroundTint="@color/cardview_light_background"
             android:layout_marginEnd="16dp"
             android:layout_marginBottom="16dp"
             android:src="@drawable/plus"
