+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,7 @@
 <resources>
     <string name="app_name">MelateKotlin</string>
     <string name="txt_mis_sorteos">Mis sorteos</string>
+    <string name="txt_button_mis_favs">Mis favoritos</string>
     <string-array name="numbers">
         <item>1</item>
         <item>2</item>
