+++ build.gradle
@@ -13,6 +13,10 @@ buildscript {
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath 'com.squareup.sqldelight:gradle-plugin:1.5.3'
 
+        // Jetpack
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.5.0-alpha03"
+
+
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
